requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy==1.24.3
faiss-cpu==1.7.4
insightface==0.7.3
asyncpg==0.29.0
pydantic==2.5.0
python-dotenv==1.0.0
Pillow==10.1.0
